+++ src/aura-helper.js
@@ -15,16 +15,19 @@ Commands.Metadata.LocalPackageGenerator.createCommand(program);
 Commands.Metadata.LocalRetrieveSpecial.createCommand(program);
 // TODO: Get queriable objects (refresh metadata index) (all and specifics)
 
-
 // OrgCommands
 Commands.Metadata.OrgList.createCommand(program);
 Commands.Metadata.OrgDescribe.createCommand(program);
 Commands.Metadata.OrgCompare.createCommand(program);
 Commands.Metadata.OrgRetrieveSpecial.createCommand(program);
+Commands.Metadata.OrgPermissions.createCommand(program);
 
+// DATA COMMANDS
+Commands.Data.Export.createCommand(program);
+Commands.Data.Import.createCommand(program);
 
 // Core Commands
-//Commands.Core.Update.createCommand(program);
+Commands.Core.Update.createCommand(program);
 
 
 program.parse(process.argv);