+++ src/commands/metadata/orgRetrieveSpecial.js
@@ -7,6 +7,7 @@ const Metadata = require('../../metadata');
 const Languages = require('../../languages');
 const ProcessManager = require('../../processes').ProcessManager;
 const Utils = require('./utils');
+const CommandUtils = require('../utils');
 const Paths = FileSystem.Paths;
 const FileChecker = FileSystem.FileChecker;
 const FileReader = FileSystem.FileReader;
@@ -27,6 +28,16 @@ const SUBFOLDER_BY_METADATA_TYPE = {
 
 let retrievedFinished = false;
 
+let argsList = [
+    "root",
+    "all",
+    "type",
+    "orgNamespace",
+    "compress",
+    "progress",
+    "compress"
+];
+
 exports.createCommand = function (program) {
     program
         .command('metadata:org:retrieve:special')
@@ -36,7 +47,7 @@ exports.createCommand = function (program) {
         .option('-t, --type <MetadataTypeNames>', 'Retrieve specifics metadata types. You can choose one or a comma separated list of elements. Also you can choose retrieve a specific profile, object o record type. Schema -> "Type1" or "Type1,Type2" or "Type1:Object1, Type1:Object2" or "Type1:Object1:Item1" for example:  "Profile, PermissinSet" for retrieve all profiles and permission sets. "Profile:Admin" for retrieve the admin profile. "RecordType:Account:RecordType1" for  retrieve the RecordType1 for the object Account or "RecordType:Account" for retrieve all Record Types for Account')
         .option('-o, --org-namespace', 'Option for retrieve only the data from your org namespace or retrieve all data')
         .option('-c, --compress', 'Compress the retrieved files.')
-        .option('-p, --progress [format]', 'Option for report the command progress. Available formats: ' + Utils.getProgressAvailableTypes().join(','))
+        .option('-p, --progress <format>', 'Option for report the command progress. Available formats: ' + CommandUtils.getProgressAvailableTypes().join(','))
         .option('-b, --beautify', 'Option for draw the output with colors. Green for Successfull, Blue for progress, Yellow for Warnings and Red for Errors. Only recomended for work with terminals (CMD, Bash, Power Shell...)')
         .action(function (args) {
             run(args);
@@ -45,7 +56,7 @@ exports.createCommand = function (program) {
 
 async function run(args) {
     Output.Printer.setColorized(args.beautify);
-    if (hasEmptyArgs(args)) {
+    if (CommandUtils.hasEmptyArgs(args, argsList)) {
         Output.Printer.printError(Response.error(ErrorCodes.MISSING_ARGUMENTS));
         return;
     }
@@ -64,8 +75,8 @@ async function run(args) {
         return;
     }
     if (args.progress) {
-        if (!Utils.getProgressAvailableTypes().includes(args.progress)) {
-            Output.Printer.printError(Response.error(ErrorCodes.MISSING_ARGUMENTS, "Wrong --progress value. Please, select any  of this vales: " + Utils.getProgressAvailableTypes().join(',')));
+        if (!CommandUtils.getProgressAvailableTypes().includes(args.progress)) {
+            Output.Printer.printError(Response.error(ErrorCodes.MISSING_ARGUMENTS, "Wrong --progress value. Please, select any  of this vales: " + CommandUtils.getProgressAvailableTypes().join(',')));
             return;
         }
     }
@@ -88,10 +99,6 @@ async function run(args) {
         retrievedFinished = true;
         Output.Printer.printError(Response.error(ErrorCodes.METADATA_ERROR, error));
     });
-}
-
-function hasEmptyArgs(args) {
-    return args.root === undefined && args.all === undefined && args.orgNamespace === undefined && args.compress === undefined && args.progress === undefined && args.beautify === undefined;
 }
 
 function retrieve(args, types) {