+++ src/metadata/customLabelsUtils.js
@@ -1,19 +1,86 @@
 const Utils = require('./utils');
+const MetadataTypes = require('./metadataTypes');
 const XMLParser = require('../languages').XMLParser;
+
+const XML_METADATA = {
+    fullName: {
+        key: "fullName",
+        label: "Full Name",
+        editable: false,
+        datatype: 'string',
+        minApi: 1,
+        maxApi: -1,
+    },
+    labels: {
+        key: "labels",
+        label: "Labels",
+        editable: true,
+        datatype: 'array',
+        minApi: 1,
+        maxApi: -1,
+        metadataType: MetadataTypes.CUSTOM_LABEL,
+        xmlData: {
+            fieldKey: "fullName",
+            sortOrder: ["fullName"],
+            fields: {
+                categories: {
+                    datatype: "string",
+                    unique: false,
+                    editable: true,
+                    default: '{!value}',
+                },
+                fullName: {
+                    datatype: "string",
+                    unique: true,
+                    editable: true,
+                    default: '{!value}',
+                },
+                language: {
+                    datatype: "string",
+                    unique: false,
+                    editable: true,
+                    default: '{!value}',
+                },
+                protected: {
+                    datatype: "boolean",
+                    unique: false,
+                    editable: true,
+                    default: false,
+                },
+                shortDescription: {
+                    datatype: "string",
+                    unique: false,
+                    editable: true,
+                    default: '{!value}',
+                },
+                value: {
+                    datatype: "string",
+                    unique: false,
+                    editable: true,
+                    default: '{!value}',
+                }
+            }
+        }
+    }
+}
 
 class CustomLabelsUtils {
 
     static createCustomLabels(customLabels) {
-        let newCustomLabels;
+        let result = {};
         if (customLabels) {
-            newCustomLabels = Utils.prepareXML(customLabels, CustomLabelsUtils.createCustomLabels());
+            result = CustomLabelsUtils.createCustomLabels();
+            result = Utils.prepareXML(customLabels, result);
+            Object.keys(result).forEach(function (elementKey) {
+                if (Array.isArray(result[elementKey])) {
+                    let elementData = XML_METADATA[elementKey];
+                    Utils.sort(result[elementKey], elementData.xmlData.sortOrder);
+                }
+            });
         } else {
-            newCustomLabels = {
-                fullName: undefined,
-                labels: []
-            }
+            result = Utils.createXMLFile(XML_METADATA);
         }
-        return newCustomLabels;
+        return result;
     }
 
     static createCustomLabel(categories, fullName, language, isProtected, shortDescription, value) {